Development/EcommerceMicroServices/product-service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart 
python-jose[cryptography]
